LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/test_no_previous_document.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>window.performance.timing attributes on an initial navigation</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
	<link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-navigation-timing-interface"/>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script id="metadata_cache">/*
{
  "window.performance is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  },
  "window.performance.navigation is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  },
  "timing.navigation.type is TYPE_NAVIGATE": { "help": "http://www.w3.org/TR/navigation-timing/#sec-navigation-info-interface" },
  "timing.unloadEventStart == 0 on navigation with no previous document": {},
  "timing.unloadEventEnd == 0 navigation with no previous document": {}
}
*/</script>
        <link rel="stylesheet" href="/w3c/resources/testharness.css" />
    
        <script>
            setup({explicit_done: true});

            var frame;
            function onload_test()
            {
                frame = document.getElementById("frameContext");
                test_namespace('navigation');
                if (performanceNamespace)
                {
                    test_true(frame.contentWindow.performance.navigation.type == performanceNamespace.navigation.TYPE_NAVIGATE,
                              'timing.navigation.type is TYPE_NAVIGATE',{help:"http://www.w3.org/TR/navigation-timing/#sec-navigation-info-interface"});

                    test_equals(frame.contentWindow.performance.timing.unloadEventStart, 0, 'timing.unloadEventStart == 0 on navigation with no previous document');
                    test_equals(frame.contentWindow.performance.timing.unloadEventEnd, 0, 'timing.unloadEventEnd == 0 navigation with no previous document');
                }
                done();
            }
        </script>
        
    </head>
    <body onload="onload_test();">
        <h1>Description</h1>
        <p>This test validates the unload event times are 0 when there is no previous document.</p>

        <div id="log"></div><br />
        <iframe id="frameContext" src="/w3c/webperf/resources/blank_page_green.htm" style="width: 250px; height: 250px;"></iframe>
    </body>
</html>
